Added `subprocess` package and moved kill function to it.
