+++ backend/ventserver/integration/_trio.py
@@ -3,7 +3,6 @@
 import functools
 import logging
 import time
-import sys
 from typing import Callable, Optional, TypeVar, Tuple, List, Type
 
 import trio
@@ -27,26 +26,6 @@ logger = logging.getLogger(__name__)
 
 
 _InputEvent = TypeVar('_InputEvent')
-
-
-# Kill frontend
-
-async def kill_frozen_frontend(
-    websocket_connected: bool, websocket_connection_time: float
-) -> None:
-    """Spawns subprocess to kill the frontend"""
-    connection_duration = int(time.time() - websocket_connection_time)
-    if websocket_connected and connection_duration > 2:
-        try:
-            print("kill called")
-            _ = await trio.run_process(
-                ["killall", "/usr/lib/chromium-browser/chromium-browser-v7"]
-            )
-            print(_)
-            logger.info("No message received from frontend for more "
-                "than a 1s; killed frontend process.")
-        except OSError as exc:
-            logger.warning("Failed to kill the frontend: %s", exc)
 
 
 # Protocol send outputs
@@ -88,10 +67,10 @@ async def send_all_websocket(
     """
     for send_event in send_channel.output_all():
         if not websocket.is_open:
-#             logger.warning(
-#                 'Discarding because websocket I/O endpoint is not open: %s',
-#                 send_event
-#             )
+            logger.warning(
+                'Discarding because websocket I/O endpoint is not open: %s',
+                send_event
+            )
             await trio.sleep(0)
             continue
