+++ backend/ventserver/io/trio/fileio.py
@@ -57,7 +57,7 @@ class Handler(endpoints.IOEndpoint[bytes, bytes]):
         except OSError as err:
             raise OSError("Handler: {}".format(err)) from err
 
-#         self._logger.info('File %s opened.', self.props.filename)
+        self._logger.info('File %s opened.', self.props.filename)
         self._connected.set()
 
     @property
@@ -73,7 +73,7 @@ class Handler(endpoints.IOEndpoint[bytes, bytes]):
         assert self._fileobject is not None
         await self._fileobject.aclose()
         self._fileobject = None
-#         self._logger.info('File %s closed.', self.props.filename)
+        self._logger.info('File %s closed.', self.props.filename)
         self._connected = trio.Event()
 
     async def receive(self) -> bytes:
