+++ backend/ventserver/simulation.py
@@ -21,6 +21,7 @@ from ventserver.integration import _trio
 from ventserver.io.trio import channels
 from ventserver.io.trio import websocket
 from ventserver.io.trio import fileio
+from ventserver.io.subprocess import frozen_frontend
 from ventserver.protocols import server
 from ventserver.protocols import exceptions
 from ventserver.protocols.protobuf import mcu_pb
@@ -360,10 +361,10 @@ async def main() -> None:
     channel: channels.TrioChannel[
         server.ReceiveOutputEvent
     ] = channels.TrioChannel()
-    
-    # Kill frontend
+
+    # Kill frozen frontend
     last_kill_call = trio.current_time()
-    
+
     # Initialize States
     states: List[Type[betterproto.Message]] = [
         mcu_pb.Parameters, mcu_pb.CycleMeasurements,
@@ -405,10 +406,11 @@ async def main() -> None:
                         if int(current_time - last_kill_call) > 2:
                             last_kill_call = current_time
                             nursery.start_soon(
-                                _trio.kill_frozen_frontend,
+                                frozen_frontend.kill_frozen_frontend,
                                 websocket_endpoint.is_open,
                                 websocket_endpoint.connection_time
                             )
+
                 nursery.cancel_scope.cancel()
     except trio.EndOfChannel:
         print('Finished, quitting!')
